src/requirements.txt
notion~=0.0.28
#jupyter
requests~=2.28.1
#PyYAML~=6.0
#pandas~=1.5.0
django-qrcode~=0.3
#matplotlib~=3.6.0
Pillow~=9.2.0
qrcode